requirements.txt
aiogram==2.25.2
aiohappyeyeballs==2.6.1
aiohttp==3.8.6
aiosignal==1.4.0
async-timeout==4.0.3
attrs==25.4.0
Babel==2.9.1
certifi==2025.10.5
charset-normalizer==3.4.3
decorator==5.2.1
frozenlist==1.8.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
instaloader==4.14.2
magic-filter==1.0.12
moviepy==2.2.1
multidict==6.7.0
nodejs-wheel-binaries==22.20.0
numpy==2.3.3
pillow==11.3.0
proglog==0.1.12
propcache==0.4.1
python-dotenv==1.1.1
pytubefix==10.0.0
pytz==2025.2
requests==2.32.5
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.22.0
yt-dlp==2025.9.26
